1_git_achieve_etl/src/requirements.txt
numpy == 1.21.6
pandas == 1.4.2
google-cloud-bigquery
db-dtypes
requests
bs4
python-dotenv